prettify a bit
